Celery work and beat has been configured
