--- tasks.py
@@ -1,6 +1,0 @@
-from celery_app import app
-
-
-@app.task
-def add(x, y):
-    return x + y
